requirements.txt
llama-cpp-python==0.2.89
huggingface-hub
fastapi
pydantic
uvicorn
datasets